requirements.txt
behave
selenium
python-dotenv